Split out the write part of the GitHub API client
This makes it clear that diffs only perform writes and no reads. It should also make mocking of writes easier (to be added later).
