+++ src/github/mod.rs
@@ -7,7 +7,7 @@ use rust_team_data::v1::Bot;
 use std::collections::{HashMap, HashSet};
 use std::fmt::Write;
 
-pub use self::api::GitHub;
+pub(crate) use self::api::{GitHub, GitHubWrite};
 
 static DEFAULT_DESCRIPTION: &str = "Managed by the rust-lang/team repository.";
 static DEFAULT_PRIVACY: TeamPrivacy = TeamPrivacy::Closed;
@@ -479,7 +479,7 @@ pub(crate) struct Diff {
 
 impl Diff {
     /// Apply the diff to GitHub
-    pub(crate) fn apply(self, sync: &SyncGitHub) -> anyhow::Result<()> {
+    pub(crate) fn apply(self, sync: &GitHubWrite) -> anyhow::Result<()> {
         for team_diff in self.team_diffs {
             team_diff.apply(sync)?;
         }
@@ -511,7 +511,7 @@ enum RepoDiff {
 }
 
 impl RepoDiff {
-    fn apply(&self, sync: &SyncGitHub) -> anyhow::Result<()> {
+    fn apply(&self, sync: &GitHubWrite) -> anyhow::Result<()> {
         match self {
             RepoDiff::Create(c) => c.apply(sync),
             RepoDiff::Update(u) => u.apply(sync),
@@ -537,10 +537,8 @@ struct CreateRepoDiff {
 }
 
 impl CreateRepoDiff {
-    fn apply(&self, sync: &SyncGitHub) -> anyhow::Result<()> {
-        let repo = sync
-            .github
-            .create_repo(&self.org, &self.name, &self.description)?;
+    fn apply(&self, sync: &GitHubWrite) -> anyhow::Result<()> {
+        let repo = sync.create_repo(&self.org, &self.name, &self.description)?;
 
         for permission in &self.permissions {
             permission.apply(sync, &self.org, &self.name)?;
@@ -592,13 +590,14 @@ impl UpdateRepoDiff {
             && self.branch_protection_diffs.is_empty()
     }
 
-    fn apply(&self, sync: &SyncGitHub) -> anyhow::Result<()> {
+    fn apply(&self, sync: &GitHubWrite) -> anyhow::Result<()> {
         if let Some((_, description)) = &self.description_diff {
-            sync.github.edit_repo(&self.org, &self.name, description)?;
+            sync.edit_repo(&self.org, &self.name, description)?;
         }
         for permission in &self.permission_diffs {
             permission.apply(sync, &self.org, &self.name)?;
         }
+
         for branch_protection in &self.branch_protection_diffs {
             branch_protection.apply(sync, &self.org, &self.name, &self.repo_id)?;
         }
@@ -640,25 +639,25 @@ struct RepoPermissionAssignmentDiff {
 }
 
 impl RepoPermissionAssignmentDiff {
-    fn apply(&self, sync: &SyncGitHub, org: &str, repo_name: &str) -> anyhow::Result<()> {
+    fn apply(&self, sync: &GitHubWrite, org: &str, repo_name: &str) -> anyhow::Result<()> {
         match &self.diff {
             RepoPermissionDiff::Create(p) | RepoPermissionDiff::Update(_, p) => {
                 match &self.collaborator {
-                    RepoCollaborator::Team(team_name) => sync
-                        .github
-                        .update_team_repo_permissions(org, repo_name, team_name, p)?,
-                    RepoCollaborator::User(user_name) => sync
-                        .github
-                        .update_user_repo_permissions(org, repo_name, user_name, p)?,
+                    RepoCollaborator::Team(team_name) => {
+                        sync.update_team_repo_permissions(org, repo_name, team_name, p)?
+                    }
+                    RepoCollaborator::User(user_name) => {
+                        sync.update_user_repo_permissions(org, repo_name, user_name, p)?
+                    }
                 }
             }
             RepoPermissionDiff::Delete(_) => match &self.collaborator {
-                RepoCollaborator::Team(team_name) => sync
-                    .github
-                    .remove_team_from_repo(org, repo_name, team_name)?,
-                RepoCollaborator::User(user_name) => sync
-                    .github
-                    .remove_collaborator_from_repo(org, repo_name, user_name)?,
+                RepoCollaborator::Team(team_name) => {
+                    sync.remove_team_from_repo(org, repo_name, team_name)?
+                }
+                RepoCollaborator::User(user_name) => {
+                    sync.remove_collaborator_from_repo(org, repo_name, user_name)?
+                }
             },
         }
         Ok(())
@@ -705,21 +704,21 @@ struct BranchProtectionDiff {
 impl BranchProtectionDiff {
     fn apply(
         &self,
-        sync: &SyncGitHub,
+        sync: &GitHubWrite,
         org: &str,
         repo_name: &str,
         repo_id: &str,
     ) -> anyhow::Result<()> {
         match &self.operation {
             BranchProtectionDiffOperation::Create(bp) => {
-                sync.github.upsert_branch_protection(
+                sync.upsert_branch_protection(
                     BranchProtectionOp::CreateForRepo(repo_id.to_string()),
                     &self.pattern,
                     bp,
                 )?;
             }
             BranchProtectionDiffOperation::Update(id, _, bp) => {
-                sync.github.upsert_branch_protection(
+                sync.upsert_branch_protection(
                     BranchProtectionOp::UpdateBranchProtection(id.clone()),
                     &self.pattern,
                     bp,
@@ -731,7 +730,7 @@ impl BranchProtectionDiff {
                 the protection is not in the team repo",
                     self.pattern, org, repo_name
                 );
-                sync.github.delete_branch_protection(org, repo_name, id)?;
+                sync.delete_branch_protection(org, repo_name, id)?;
             }
         }
 
@@ -799,7 +798,7 @@ enum TeamDiff {
 }
 
 impl TeamDiff {
-    fn apply(self, sync: &SyncGitHub) -> anyhow::Result<()> {
+    fn apply(self, sync: &GitHubWrite) -> anyhow::Result<()> {
         match self {
             TeamDiff::Create(c) => c.apply(sync)?,
             TeamDiff::Edit(e) => e.apply(sync)?,
@@ -829,9 +828,8 @@ struct CreateTeamDiff {
 }
 
 impl CreateTeamDiff {
-    fn apply(self, sync: &SyncGitHub) -> anyhow::Result<()> {
-        sync.github
-            .create_team(&self.org, &self.name, &self.description, self.privacy)?;
+    fn apply(self, sync: &GitHubWrite) -> anyhow::Result<()> {
+        sync.create_team(&self.org, &self.name, &self.description, self.privacy)?;
         for (member_name, role) in self.members {
             MemberDiff::Create(role).apply(&self.org, &self.name, &member_name, sync)?;
         }
@@ -872,12 +870,12 @@ struct EditTeamDiff {
 }
 
 impl EditTeamDiff {
-    fn apply(self, sync: &SyncGitHub) -> anyhow::Result<()> {
+    fn apply(self, sync: &GitHubWrite) -> anyhow::Result<()> {
         if self.name_diff.is_some()
             || self.description_diff.is_some()
             || self.privacy_diff.is_some()
         {
-            sync.github.edit_team(
+            sync.edit_team(
                 &self.org,
                 &self.name,
                 self.name_diff.as_deref(),
@@ -946,12 +944,12 @@ enum MemberDiff {
 }
 
 impl MemberDiff {
-    fn apply(self, org: &str, team: &str, member: &str, sync: &SyncGitHub) -> anyhow::Result<()> {
+    fn apply(self, org: &str, team: &str, member: &str, sync: &GitHubWrite) -> anyhow::Result<()> {
         match self {
             MemberDiff::Create(role) | MemberDiff::ChangeRole((_, role)) => {
-                sync.github.set_team_membership(org, team, member, role)?;
+                sync.set_team_membership(org, team, member, role)?;
             }
-            MemberDiff::Delete => sync.github.remove_team_membership(org, team, member)?,
+            MemberDiff::Delete => sync.remove_team_membership(org, team, member)?,
             MemberDiff::Noop => {}
         }
 
@@ -970,8 +968,8 @@ struct DeleteTeamDiff {
 }
 
 impl DeleteTeamDiff {
-    fn apply(self, sync: &SyncGitHub) -> anyhow::Result<()> {
-        sync.github.delete_team(&self.org, &self.slug)?;
+    fn apply(self, sync: &GitHubWrite) -> anyhow::Result<()> {
+        sync.delete_team(&self.org, &self.slug)?;
         Ok(())
     }
 }
